+++ app/Component/Column/TemplateColumn.js
