+++ app/Component/Column/index.js
@@ -1,0 +1,15 @@
+import BooleanColumn from './BooleanColumn';
+import DateColumn from './DateColumn';
+import NumberColumn from './NumberColumn';
+import ReferenceColumn from './ReferenceColumn';
+import ReferenceManyColumn from './ReferenceManyColumn';
+import TemplateColumn from './TemplateColumn';
+
+export {
+    BooleanColumn,
+    DateColumn,
+    NumberColumn,
+    ReferenceColumn,
+    ReferenceManyColumn,
+    TemplateColumn
+};
