--- app/Component/Datagrid/Field/index.js
@@ -1,15 +1,0 @@
-import BooleanField from './BooleanField';
-import DateField from './DateField';
-import NumberField from './NumberField';
-import ReferenceField from './ReferenceField';
-import ReferenceManyField from './ReferenceManyField';
-import TemplateField from './TemplateField';
-
-export {
-    BooleanField,
-    DateField,
-    NumberField,
-    ReferenceField,
-    ReferenceManyField,
-    TemplateField
-};
